shifts tests to correct testing pages.
